requirements.txt
face_alignment
kornia
pillow
torch
torchvision
tqdm
